Factor HIV1B-specific procedures out of core CLI
In preparation for adding HCV support, this commit factors
HIV1B-specific features out of the core CLI procedure. This leaves the
HIV1B specific configuration data in one module, with the the bulk of
the common operations (IO, concurrency, argument wrangling, etc.) in a
separate procedure.

The resulting "architecture" is that the `cli/hiv1b` module (and other
modules in the future) expose a `PerformAlignment` function that might
perform some configuration or argument checking, but mostly pulls in
extra information (such as reference sequences) and hands off to the
core `PerformAlignment` function in `cli/cli`.

To make passing arguments through the genome-specific
`PerformAlignment` function less verbose, this commit also adds to
structs for storing arguments, one for input and output arguments, and
one for generic alignment parameters.