Implement a sort interface on []rawIndelScore
This lets us sort raw indel scores before we serialize them. If we
didn't do this, when we deserialize a profile's positional indel
scores the resulting profile would have the indel scores in
whatever (unpredictable) order they came out the positional indel
scores map (which is un-ordered). By sorting the output before we
serialize it, we can ensure that the resulting profiles are
consistent.