Add align command
This commit adds the body of the `nucamino align` command. It takes
two arguments, an alignment profile name (e.g. hiv1b) and a comma
separated list of genes from that profile to include in the
alignment. It has flags for other parameters that have sensible
defaults (e.g. goroutines, input and output files).

This commit also includes some changes to the `cli` package:

- Remove the IOParameters struct; it's not needed anymore, as we no
  longer pass these arguments around as much.
- Check that the output format is valid before we run the alignment.
- Change the return type of PerformAlignment from void to error, so
  that we can propagate errors (such as missing files or invalid
  output types) up to cobra to be displayed to the user in a
  consistent manner.