Add align-with command
This is similar to the align command, but loads its alignment profile
from a YAML file instead of using the built-in profiles. Code was
duplicated between them as their APIs may diverge, and it's easier to
track two separate implementations of the same CLI for a little while
than it is to separate prematurely abstracted code.