Add a command to list an alignment profile's genes
This commit adds a subcommand (`nucamino profile list-genes`) that
lists the genes available in built-in alignment profile. Like
`nucamino align` it takes a profile name as a required first argument
and errors if it doesn't match one of the built-in profiles. Like
`nucamino profile list` it takes an optional regular expression to
filter the output.

Example:

    $ nucamino profile list-genes hcv1a
    NS5B
    NS3
    NS5A

    $ nucamino profile list-genes hiv1b ^G
    GAG
    GP41